feat(tests): add TypeScript configuration for test files and refactor test cases
